novos testes e validação no PersistTransactionUseCase

+++ 05-spring-ai/src/main/java/dio/budgeting/infrastructure/http/response/TransactionResponse.java
@@ -2,7 +2,9 @@ package dio.budgeting.infrastructure.http.response;
 
 import dio.budgeting.application.output.TransactionOutput;
 
-public record TransactionResponse(String id, String category, String description, double amount) {
+import java.math.BigDecimal;
+
+public record TransactionResponse(String id, String category, String description, BigDecimal amount) {
     public static TransactionResponse from(TransactionOutput output) {
         return new TransactionResponse(output.id(), output.category(), output.description(), output.value());
     }
